+++ examples/event-stream-smol.rs
@@ -1,10 +1,10 @@
-//! Demonstrates how to read events asynchronously with async-std.
+//! Demonstrates how to read events asynchronously with smol.
 //!
-//! cargo run --features="event-stream" --example event-stream-async-std
+//! cargo run --features="event-stream" --example event-stream-smol
 
 use std::{io::stdout, time::Duration};
 
-use futures::{future::FutureExt, select, StreamExt};
+use futures::{StreamExt, future::FutureExt, select};
 use futures_timer::Delay;
 
 use crossterm::{
@@ -14,7 +14,7 @@ use crossterm::{
     terminal::{disable_raw_mode, enable_raw_mode},
 };
 
-const HELP: &str = r#"EventStream based on futures_util::stream::Stream with async-std
+const HELP: &str = r#"EventStream based on futures_util::stream::Stream with smol
  - Keyboard, mouse and terminal resize events enabled
  - Prints "." every second if there's no event
  - Hit "c" to print current cursor position
@@ -59,7 +59,7 @@ fn main() -> std::io::Result<()> {
     let mut stdout = stdout();
     execute!(stdout, EnableMouseCapture)?;
 
-    async_std::task::block_on(print_events());
+    smol::block_on(print_events());
 
     execute!(stdout, DisableMouseCapture)?;