+++ src/event.rs
@@ -133,9 +133,8 @@ use derive_more::derive::IsVariant;
 pub use stream::EventStream;
 
 use crate::{
-    csi,
+    Command, csi,
     event::{filter::EventFilter, internal::InternalEvent},
-    Command,
 };
 use std::fmt::{self, Display};
 use std::time::Duration;