+++ src/event/read.rs
@@ -53,12 +53,7 @@ impl InternalEventReader {
 
         let event_source = match self.source.as_mut() {
             Some(source) => source,
-            None => {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::Other,
-                    "Failed to initialize input reader",
-                ))
-            }
+            None => return Err(std::io::Error::other("Failed to initialize input reader")),
         };
 
         let poll_timeout = PollTimeout::new(timeout);
@@ -171,12 +166,16 @@ mod tests {
         };
 
         assert!(reader.poll(None, &InternalEventFilter).is_err());
-        assert!(reader
-            .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
-            .is_err());
-        assert!(reader
-            .poll(Some(Duration::from_secs(10)), &InternalEventFilter)
-            .is_err());
+        assert!(
+            reader
+                .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
+                .is_err()
+        );
+        assert!(
+            reader
+                .poll(Some(Duration::from_secs(10)), &InternalEventFilter)
+                .is_err()
+        );
     }
 
     #[test]
@@ -277,9 +276,11 @@ mod tests {
             skipped_events: Vec::with_capacity(32),
         };
 
-        assert!(!reader
-            .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
-            .unwrap());
+        assert!(
+            !reader
+                .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
+                .unwrap()
+        );
     }
 
     #[test]
@@ -293,9 +294,11 @@ mod tests {
         };
 
         assert!(reader.poll(None, &InternalEventFilter).unwrap());
-        assert!(reader
-            .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
-            .unwrap());
+        assert!(
+            reader
+                .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
+                .unwrap()
+        );
     }
 
     #[test]
@@ -345,9 +348,11 @@ mod tests {
         assert_eq!(reader.read(&InternalEventFilter).unwrap(), EVENT);
         assert_eq!(reader.read(&InternalEventFilter).unwrap(), EVENT);
         assert_eq!(reader.read(&InternalEventFilter).unwrap(), EVENT);
-        assert!(!reader
-            .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
-            .unwrap());
+        assert!(
+            !reader
+                .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
+                .unwrap()
+        );
     }
 
     #[test]
@@ -398,9 +403,11 @@ mod tests {
 
         assert_eq!(reader.read(&InternalEventFilter).unwrap(), EVENT);
         assert!(reader.read(&InternalEventFilter).is_err());
-        assert!(reader
-            .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
-            .unwrap());
+        assert!(
+            reader
+                .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
+                .unwrap()
+        );
     }
 
     #[test]
@@ -430,7 +437,7 @@ mod tests {
         fn new(events: &[InternalEvent]) -> FakeSource {
             FakeSource {
                 events: events.to_vec().into(),
-                error: Some(io::Error::new(io::ErrorKind::Other, "")),
+                error: Some(io::Error::other("")),
             }
         }