+++ src/event/sys/windows.rs
@@ -37,7 +37,7 @@ fn init_original_console_mode(original_mode: u32) {
 /// Returns the original console color, make sure to call `init_console_color` before calling this function. Otherwise this function will panic.
 pub(crate) fn original_console_mode() -> std::io::Result<u32> {
     u32::try_from(ORIGINAL_CONSOLE_MODE.load(Ordering::Relaxed))
-        .map_err(|_| io::Error::new(io::ErrorKind::Other, "Initial console modes not set"))
+        .map_err(|_| io::Error::other("Initial console modes not set"))
 }
 
 /// Try to enable virtual terminal input on the console input handle.